Made schema conversion functions public so they can be used outside storage
